refactor some shitty code before anything else
