+++ frontend/src/components/formGenerator/formGenerator.scss
@@ -1,0 +1,79 @@
+@import '../../variables.scss';
+
+.form {
+  @include flex(column, center, center);
+  gap: $gap-medium;
+  margin: $spacing-small;
+  background-color: $transparent-black;
+  border-radius: $border-radius-20;
+  padding: $spacing-medium;
+  width: 500px;
+
+  @media (max-width: 768px) {
+    width: 90%;
+  }
+
+  &__title {
+    font-size: $fs-large;
+    font-family: $font-fam-primary;
+    letter-spacing: $spacing-small;
+    text-transform: uppercase;
+  }
+
+  &__group {
+    @include flex(column, center, center);
+    gap: $gap-small;
+    margin-bottom: $spacing-small;
+    padding: $spacing-small;
+    width: 80%;
+
+    &__input {
+      text-align: center;
+      height: $spacing-large;
+      border: $border-1 $primary-green;
+      border-radius: $border-radius-10;
+      font-family: $font-fam-primary;
+      font-weight: $weight-600;
+      background-color: $primary-black;
+      color: $primary-white;
+      width: 70%;
+    }
+  }
+
+  &__btn {
+    cursor: pointer;
+    background-color: $primary-green;
+    font-family: $font-fam-primary;
+    padding: $spacing-small;
+    font-size: $fs-medium;
+    font-weight: $weight-600;
+    border: $border-0;
+    border-radius: $border-radius-10;
+    text-transform: uppercase;
+    width: 70%;
+  }
+
+  &__redirect {
+    cursor: pointer;
+    background-color: $primary-green;
+    font-family: $font-fam-primary;
+    padding: $spacing-small;
+    font-size: $fs-medium;
+    font-weight: $weight-600;
+    border: $border-0;
+    border-radius: $border-radius-10;
+    text-transform: uppercase;
+    width: 500px;
+
+    @media (max-width: 768px) {
+      width: 90%;
+      font-size: $fs-xsmall;
+    }
+
+    &__link {
+      text-decoration: none;
+      color: $primary-white;
+      font-size: $fs-small;
+    }
+  }
+}
