+++ frontend/src/components/formGenerator/formGenerator.tsx
@@ -1,11 +1,9 @@
 import { useForm } from 'react-hook-form'
-import '../register/register.scss'
-import '../login/login.scss'
+import './formGenerator.scss'
 
 interface FormProps {
   onSubmit: (data: any) => void
   inputConfigs?: any
-  configsLogin?: any
 }
 
 const FormGenerator = (props: FormProps) => {
