+++ frontend/src/pages/login/login.scss
@@ -1,87 +1,10 @@
 @import '../../variables.scss';
 
-.card {
+.form__login {
   @include flex(column, center, center);
   text-align: center;
   font-family: $font-fam-primary;
   font-weight: $weight-600;
   color: $primary-white;
-  background: url('../../assets/menu-background.gif');
-  background-repeat: no-repeat;
-  background-size: cover;
-  height: 100vh;
-
-  &__form {
-    @include flex(column, center, center);
-    gap: $gap-medium;
-    margin: $spacing-small;
-    background-color: $transparent-black;
-    border-radius: $border-radius-20;
-    padding: $spacing-medium;
-    width: 500px;
-    @include media(193px, $breakpoint-phone);
-
-    &__title {
-      font-size: $fs-large;
-      font-family: $font-fam-primary;
-      letter-spacing: $spacing-small;
-      text-transform: uppercase;
-    }
-
-    &__group {
-      @include flex(column, center, center);
-      gap: $gap-small;
-      margin-bottom: $spacing-small;
-      padding: $spacing-small;
-      width: 80%;
-      @include media(150px, $breakpoint-phone);
-
-      &__input {
-        text-align: center;
-        padding: $spacing-small $spacing-large;
-        height: $spacing-large;
-        border: $border-1 $primary-green;
-        border-radius: $border-radius-10;
-        font-family: $font-fam-primary;
-        font-weight: $weight-600;
-        background-color: $primary-black;
-        color: $primary-white;
-        width: 70%;
-        @include media(150px, $breakpoint-phone);
-      }
-    }
-
-    &__btn {
-      cursor: pointer;
-      background-color: $primary-green;
-      font-family: $font-fam-primary;
-      padding: $spacing-small $spacing-large;
-      font-size: $fs-medium;
-      font-weight: $weight-600;
-      border: $border-0;
-      border-radius: $border-radius-10;
-      text-transform: uppercase;
-      width: 70%;
-      @include media(150px, $breakpoint-phone);
-    }
-
-    &__redirect {
-      cursor: pointer;
-      background-color: $primary-green;
-      font-family: $font-fam-primary;
-      padding: $spacing-small;
-      font-size: $fs-medium;
-      font-weight: $weight-600;
-      border: $border-0;
-      border-radius: $border-radius-10;
-      text-transform: uppercase;
-      @include media(193px, $breakpoint-phone);
-
-      &__link {
-        text-decoration: none;
-        color: $primary-white;
-        font-size: $fs-small;
-      }
-    }
-  }
+  padding-top: 50px;
 }
